pyproject.toml
[project]
name = "project-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "hadjuse",email = "rabearimananahadjpro@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain-mistralai (>=0.2.4,<0.3.0)",
    "streamlit (>=1.41.1,<2.0.0)",
    "langchain-openai (>=0.3.2,<0.4.0)",
    "langchain (>=0.3.16,<0.4.0)",
    "langgraph (>=0.2.68,<0.3.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "ipython (>=8.31.0,<9.0.0)",
    "langsmith (>=0.3.2,<0.4.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "pypdf (>=5.2.0,<6.0.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "pymupdf (>=1.25.2,<2.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pysqlite3-binary (>=0.5.4,<0.6.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
